fix: move projectEditUseCase to project package
