+++ src/test/kotlin/logic/usecase/task/EditProjectUseCaseTest.kt
@@ -4,7 +4,7 @@ package logic.usecase.task
 import io.mockk.every
 import io.mockk.mockk
 import junit.framework.TestCase.assertTrue
-import logic.usecases.EditProjectUseCase
+import logic.usecase.project.EditProjectUseCase
 import org.example.logic.entity.Project
 import org.example.logic.repository.ProjectRepository
 import org.junit.jupiter.api.BeforeEach
